src/requirements.txt
pydantic-ai
pydantic-ai-slim[openai]
python-dotenv
logfire
httpx
openai
